generate the common state tests via macro
