+++ execution-plane/arcon/arcon_state/src/lib.rs
@@ -16,8 +16,15 @@ use std::{
 
 pub mod error;
 pub mod handles;
+#[macro_use]
+pub mod macros;
 pub mod ops;
 pub mod serialization;
+
+#[cfg(test)]
+#[macro_use]
+pub mod test_common;
+
 pub use crate::{
     handles::Handle,
     ops::{AggregatorOps, MapOps, ReducerOps, ValueOps, VecOps},
@@ -283,144 +290,6 @@ impl<A: Aggregator> Default for AggregatorState<A> {
     fn default() -> Self {
         AggregatorState(Default::default())
     }
-}
-
-/// Macro to implement [`Bundle`]s.
-///
-/// [`Bundle`]: trait.Bundle.html
-///
-/// # Examples
-///
-/// ```
-/// # #[macro_use] extern crate arcon_state;
-/// # use arcon_state::*;
-/// # use arcon_state::in_memory::*;
-/// # let mut backend = InMemory::restore_or_create(&Default::default(), Default::default()).unwrap();
-/// # let mut backend_session = backend.session();
-/// bundle! {
-///     /// My test bundle
-///     pub struct MyTestBundle<T: Value> {
-///         /// some value
-///         value: Handle<ValueState<T>>,
-///         /// some map
-///         map: Handle<MapState<String, f64>, i32, u32>
-///     }
-/// }
-///
-/// impl<T: Value> MyTestBundle<T> {
-///     fn new() -> Self {
-///         MyTestBundle {
-///             value: Handle::value("value"),
-///             map: Handle::map("map").with_item_key(-1).with_namespace(0)
-///         }       
-///     }   
-/// }
-///
-/// let mut bundle = MyTestBundle::<u32>::new();
-/// // Usually a bundle should be registered by an arcon node, but we'll do it manually here
-/// bundle.register_states(&mut backend_session, unsafe { &RegistrationToken::new() });
-///
-/// let mut active_bundle = bundle.activate(&mut backend_session);
-/// let mut value_handle = active_bundle.value();
-/// value_handle.set(3).unwrap();
-/// println!("value is {:?}", value_handle.get().unwrap());
-/// let mut map_handle = active_bundle.map();
-/// map_handle.insert("foo".into(), 0.5).unwrap();
-/// println!("map[foo] is {:?}", map_handle.get(&"foo".into()).unwrap())
-/// ```
-/// Note that if you want to have more bounds on your generic parameter, you'll have to use slightly
-/// different syntax than what you're used to from regular Rust:
-/// ```ignore
-/// pub struct ParamWithManyBounds<T: Default: Clone> { // instead of T: Default + Clone
-/// ```
-/// Also, where clauses are unsupported.
-// this might be nicer as a derive macro, but IntelliJ Rust can see through macro_rules!, so this
-// version is nicer for people that use that
-#[macro_export]
-macro_rules! bundle {
-    (
-        $(#[$bundle_meta:meta])*
-        $vis:vis struct $name:ident $(<
-            $($generic_lifetime_param:lifetime),*$(,)?
-            $($generic_param:ident $(: $first_bound:path $(: $other_bounds:path)*)?),*$(,)?
-        >)? {$(
-            $(#[$state_meta:meta])*
-            $state_name:ident : Handle<$state_type:ty $(, $item_key_type:ty $(, $namespace_type:ty)?)?>
-        ),*$(,)?}
-    ) => {
-        $(#[$bundle_meta])*
-        $vis struct $name$(<
-            $($generic_lifetime_param,)*
-            $($generic_param $(: $first_bound $(+ $other_bounds)*)?,)*
-        >)? {
-            $(
-                $(#[$state_meta])*
-                $state_name : $crate::Handle<$state_type $(, $item_key_type $(, $namespace_type)?)?>,
-            )*
-        }
-
-        const _: () = {
-            pub struct Active<
-                '__bundle, '__backend, $($($generic_lifetime_param,)*)?
-                __B: $crate::Backend,
-                $($($generic_param $(: $first_bound $(+ $other_bounds)*)?,)*)?
-            > {
-                backend: &'__backend mut __B,
-                inner: &'__bundle mut $name$(<
-                    $($generic_lifetime_param,)*
-                    $($generic_param,)*
-                >)?,
-            }
-
-            impl<
-                '__bundle, '__backend, $($($generic_lifetime_param,)*)?
-                __B: $crate::Backend,
-                $($($generic_param $(: $first_bound $(+ $other_bounds)*)?,)*)?
-            > Active<
-                '__bundle, '__backend, $($($generic_lifetime_param,)*)?
-                __B, $($($generic_param,)*)?
-            > {$(
-                $(#[$state_meta])*
-                #[inline]
-                fn $state_name(&mut self) -> $crate::handles::ActiveHandle<__B,
-                    $state_type $(, $item_key_type $(, $namespace_type)?)?
-                > {
-                    self.inner.$state_name.activate(&mut self.backend)
-                }
-            )*}
-
-            impl<
-                '__this, '__backend, $($($generic_lifetime_param,)*)?
-                __B: $crate::Backend + '__backend,
-                $($($generic_param : '__this $(+ $first_bound $(+ $other_bounds)*)?,)*)?
-            > $crate::Bundle<'__this, '__backend, __B> for $name$(<
-                $($generic_lifetime_param,)* $($generic_param,)*
-            >)? {
-                type Active = Active<
-                    '__this, '__backend, $($($generic_lifetime_param,)*)?
-                    __B, $($($generic_param,)*)?
-                >;
-
-                fn register_states<'s>(
-                    &mut self,
-                    session: &'__backend mut $crate::Session<'s, __B>,
-                    registration_token: &$crate::RegistrationToken
-                ) {
-                    $(self.$state_name.register(session, registration_token);)*
-                }
-
-                fn activate<'s>(
-                    &'__this mut self,
-                    session: &'__backend mut $crate::Session<'s, __B>,
-                ) -> Self::Active {
-                    Active {
-                        backend: session.backend,
-                        inner: self,
-                    }
-                }
-            }
-        };
-    };
 }
 
 pub mod in_memory;
